+++ tests/testthat/helper-torchvision.R
@@ -92,6 +92,16 @@ expect_coco_model_detects_cat <- function(model, min_score = 0.25, size = c(640,
   expect_gt(scores_vec[top], min_score)
 }
 
+make_segmentation_item <- function(image_size = c(100L, 200L), num_masks = 3L) {
+  x <- torch_randn(3, image_size[1], image_size[2])
+  masks <- torch_rand(num_masks, image_size[1], image_size[2]) > 0.5
+  labels <- torch_ones(num_masks, dtype = torch_long())
+  y <- list(masks = masks, labels = labels, image_height = image_size[1], image_width = image_size[2])
+  item <- list(x = x, y = y)
+  class(item) <- c("image_with_segmentation_mask", "list")
+  item
+}
+
 # Helper to build a detection item target with boxes, labels, image metadata, area and is_crowd (as per coco dataset targets)
 make_detection_target <- function(boxes,
                         labels = NULL,
@@ -133,4 +143,3 @@ make_detection_item <- function(boxes, labels = NULL, image_size = c(100L, 200L)
   class(item) <- c("image_with_bounding_box", "list")
   item
 }
-